+++ src/Commands/Files/UndoFileCheckedOut.cs
@@ -15,7 +15,7 @@ namespace PnP.PowerShell.Commands.Files
             // Remove URL decoding from the Url as that will not work. We will encode the + character specifically, because if that is part of the filename, it needs to stay and not be decoded.
             Url = Utilities.UrlUtilities.UrlDecode(Url.Replace("+", "%2B"));
 
-            CurrentWeb.UndoCheckOutFileAsync(Url);
+            //CurrentWeb.UndoCheckOutFileAsync(Url);
         }
     }
 }